useContex Hook separated files
