+++ src/AppUseContext.js
@@ -1,61 +1,12 @@
-import React, { useContext, useState, createContext } from "react";
 import "./App.css";
-
-const globalState = {
-  title: "O título que contexto",
-  body: "O body do contexto",
-  counter: 0,
-};
-
-const GlobalContext = createContext();
-
-// eslint-disable-next-line
-const Div = ({ children }) => {
-  return (
-    <>
-      <H1 />
-      <P />
-    </>
-  );
-};
-
-// eslint-disable-next-line
-const H1 = () => {
-  const theContext = useContext(GlobalContext);
-  const {
-    contextState: { title, counter },
-  } = theContext;
-  return (
-    <h1>
-      {title} {counter}
-    </h1>
-  );
-};
-
-// eslint-disable-next-line
-const P = () => {
-  const theContext = useContext(GlobalContext);
-  const {
-    contextState: { body, counter },
-    contextState,
-    setContextState,
-  } = theContext;
-  return (
-    <p
-      onClick={() => setContextState((s) => ({ ...s, counter: s.counter + 1 }))}
-    >
-      {body}
-    </p>
-  );
-};
+import { Div } from "./components/Div";
+import { AppContext } from "./contexts/AppContext";
 
 function App() {
-  const [contextState, setContextState] = useState(globalState);
-
   return (
-    <GlobalContext.Provider value={{ contextState, setContextState }}>
+    <AppContext>
       <Div />
-    </GlobalContext.Provider>
+    </AppContext>
   );
 }
