+++ src/components/Div/index.jsx
@@ -1,0 +1,12 @@
+import { H1 } from "../H1";
+import { P } from "../P";
+
+// eslint-disable-next-line
+export const Div = ({ children }) => {
+  return (
+    <>
+      <H1 />
+      <P />
+    </>
+  );
+};
