+++ src/contexts/AppContext/data.js
@@ -1,0 +1,5 @@
+export const globalState = {
+  title: "O título que contexto",
+  body: "O body do contexto",
+  counter: 0,
+};
